Created component PetsData
