+++ src/components/PetsData/index.module.css
@@ -1,0 +1,78 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  margin: 40px 20px 0;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.title {
+  font-family: 'Manrope', sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 20px;
+  line-height: 1.35;
+  letter-spacing: 0.04em;
+
+  color: var(--color-main);
+}
+
+.addPetWrapper {
+  display: flex;
+  align-items: center;
+}
+
+.addPetTitle {
+  font-family: 'Manrope', sans-serif;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 20px;
+  line-height: 1.35;
+
+  color: var(--color-main);
+}
+
+.buttonAddPet {
+  width: 40px;
+  height: 40px;
+
+  margin-left: 15px;
+
+  border: none;
+  border-radius: 50%;
+  stroke: white;
+  background-color: var(--btn-bg-another);
+  background-image: url('../../images/svg/plus.svg');
+  cursor: pointer;
+  /* fill: currentColor; */
+  background-repeat: no-repeat;
+  background-size: 24px 24px;
+  background-position: center;
+  /* stroke: white;
+  color: white; */
+}
+
+@media screen and (min-width: 768px) {
+  .wrapper {
+    margin: 20px 0 20px 0;
+  }
+
+  .title {
+    font-size: 28px;
+    line-height: 1.36;
+  }
+}
+
+@media screen and (min-width: 1280px) {
+  .header {
+    padding-right: 16px;
+  }
+
+  .buttonAddPet {
+    margin-left: 20px;
+  }
+}
